Rename .h files to .hpp
